refactor(whatsapp): keep inbound boundary transport-private
