+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -29,7 +29,7 @@ c0f910ebfa3dbf283145fb1e3b9c016d03e853ef13e70402b09f3b9d9c2f4ab0  module/channel
 9a5aaf650f9242523bb57bdc2556c323ab64e55aa11e25e2685b73b23ee12534  module/channel-dm-policy
 fbf353eb38ae68d8ded3f2a60b432c7bb2c245d2ec7e7c9f53c6da19a0db0938  module/channel-entry-contract
 982f29a18e07228e3da82cae67d06ff38249592a29c2fd28f01f0d2016ff80d9  module/channel-feedback
-6cf2da836db49a267aed470ce6f5f55e68d8d86e1efe980cdc71349cfce30bc2  module/channel-inbound
+d645d24bcb7a5f68cc46c692ad0d1fbd19be0a99996f31e9479ce9cffce301c1  module/channel-inbound
 76bb7f531f3702c801e8fe7479e9e499f601fb361a4303afdcb45fc0da440e4b  module/channel-inbound-debounce
 4a7ada095f0f483525dcbd848fbccab26473749eba87e6a6c6e5074fd04ee1d1  module/channel-ingress-runtime
 c97dd36cdf8f83c2893c33e9430a93cd131a03d855725783ca5b545de0cf84f8  module/channel-lifecycle
