+++ extensions/whatsapp/src/inbound/message-aliases.test.ts
@@ -2,7 +2,6 @@
 import { describe, expect, expectTypeOf, it, vi } from "vitest";
 import {
   normalizeAdmittedWebInboundMessage,
-  normalizeWebInboundMessage,
   withDeprecatedWebInboundMessageFlatAliases,
 } from "./message-aliases.js";
 import type { monitorWebInbox } from "./monitor.js";
@@ -281,42 +280,33 @@ describe("WhatsApp inbound flat aliases", () => {
     expect(msg.accessControlPassed).toBe(true);
   });
 
-  it("normalizes canonical messages without admission without rebuilding admission facts", () => {
-    const normalized = normalizeWebInboundMessage({
-      event: {
-        id: "group-no-admission",
-        timestamp: 1_700_000_456,
-      },
-      payload: {
-        body: "hello group",
-      },
-      platform: {
-        chatJid: "123@g.us",
-        recipientJid: "+15550000001",
-        senderJid: "15550000002@s.whatsapp.net",
-        senderE164: "+15550000002",
-        senderName: "Alice",
-        sendComposing: vi.fn(async () => undefined),
-        reply: vi.fn(async () => createAcceptedWhatsAppSendResult("text", "reply-group")),
-        sendMedia: vi.fn(async () => createAcceptedWhatsAppSendResult("media", "media-group")),
-      },
-      from: "123@g.us",
-      conversationId: "123@g.us",
-      accountId: "work",
-      accessControlPassed: true,
-      chatType: "group",
-    });
-
-    expect(normalized.admission).toBeUndefined();
-    expect(normalized.from).toBe("123@g.us");
-    expect(normalized.conversationId).toBe("123@g.us");
-    expect(normalized.accountId).toBe("work");
-    expect(normalized.chatType).toBe("group");
-
-    normalized.conversationId = "456@g.us";
-    expect(normalized.admission).toBeUndefined();
-    expect(normalized.from).toBe("456@g.us");
-    expect(normalized.conversationId).toBe("456@g.us");
+  it("rejects canonical messages without admission at the listener boundary", () => {
+    expect(() =>
+      normalizeAdmittedWebInboundMessage({
+        event: {
+          id: "group-no-admission",
+          timestamp: 1_700_000_456,
+        },
+        payload: {
+          body: "hello group",
+        },
+        platform: {
+          chatJid: "123@g.us",
+          recipientJid: "+15550000001",
+          senderJid: "15550000002@s.whatsapp.net",
+          senderE164: "+15550000002",
+          senderName: "Alice",
+          sendComposing: vi.fn(async () => undefined),
+          reply: vi.fn(async () => createAcceptedWhatsAppSendResult("text", "reply-group")),
+          sendMedia: vi.fn(async () => createAcceptedWhatsAppSendResult("media", "media-group")),
+        },
+        from: "123@g.us",
+        conversationId: "123@g.us",
+        accountId: "work",
+        accessControlPassed: true,
+        chatType: "group",
+      }),
+    ).toThrow("WhatsApp inbound message is missing admission facts");
   });
 
   it("normalizes legacy flat messages into canonical contexts with live aliases", () => {
@@ -351,7 +341,7 @@ describe("WhatsApp inbound flat aliases", () => {
       isBatched: true,
     };
 
-    const normalized = normalizeWebInboundMessage(legacy);
+    const normalized = normalizeAdmittedWebInboundMessage(legacy);
 
     expect(normalized.event).toMatchObject({
       id: "legacy-1",
@@ -405,7 +395,7 @@ describe("WhatsApp inbound flat aliases", () => {
   });
 
   it("normalizes blocked legacy flat admission fields through the compatibility seam", () => {
-    const normalized = normalizeWebInboundMessage({
+    const normalized = normalizeAdmittedWebInboundMessage({
       id: "legacy-blocked",
       from: "+15550000002",
       conversationId: "+15550000002",
@@ -437,7 +427,7 @@ describe("WhatsApp inbound flat aliases", () => {
   });
 
   it("preserves explicit legacy access proof through normalization", () => {
-    const normalized = normalizeWebInboundMessage({
+    const normalized = normalizeAdmittedWebInboundMessage({
       ...createTestLegacyFlatWebInboundMessage(),
       accessControlPassed: true,
     });
