+++ src/plugin-sdk/channel-inbound.test.ts
@@ -5,8 +5,6 @@ import { describe, expect, expectTypeOf, it } from "vitest";
 import {
   buildChannelInboundEventContext,
   type BuildChannelInboundEventContextParams,
-  type PreparedChannelInbound,
-  projectPreparedChannelInbound,
   type PluginHookChannelSenderContext,
 } from "./channel-inbound.js";
 
@@ -72,98 +70,5 @@ describe("channel-inbound public helpers", () => {
     );
 
     expect(ctx.ChannelContext?.sender?.testUnionId).toBe("union-1");
-  });
-
-  it("builds a portable prepared inbound without channel-native types", () => {
-    const inbound = {
-      channel: "example",
-      accountId: "work",
-      event: {
-        id: "event-1",
-        fullId: "example:event-1",
-        timestamp: 1_710_000_000,
-      },
-      from: "example:user:u1",
-      sender: {
-        id: "u1",
-        name: "Alice",
-      },
-      conversation: {
-        kind: "group",
-        id: "room-1",
-        label: "Example Room",
-      },
-      route: {
-        agentId: "main",
-        accountId: "work",
-        routeSessionKey: "agent:main:example:group:room-1",
-      },
-      reply: {
-        to: "example:room:room-1",
-        replyToId: "quoted-1",
-      },
-      message: {
-        body: "agent body",
-        bodyForAgent: "agent body",
-        rawBody: "raw body",
-        commandBody: "/status",
-      },
-      command: {
-        kind: "text-slash",
-        body: "/status",
-        authorization: {
-          kind: "denied",
-          reason: "sender_not_allowed",
-        },
-      },
-      media: [
-        {
-          path: "/tmp/example.jpg",
-          contentType: "image/jpeg",
-          kind: "image",
-        },
-      ],
-      context: {
-        senderE164: "+15550001111",
-      },
-    } satisfies PreparedChannelInbound;
-
-    const projected = projectPreparedChannelInbound({
-      inbound,
-      control: { messageReceivedHooks: "core" },
-    });
-    expect(projected.input).toEqual({
-      id: "event-1",
-      timestamp: 1_710_000_000,
-      rawText: "raw body",
-      textForAgent: "agent body",
-      textForCommands: "/status",
-      raw: inbound,
-    });
-    expect(inbound.command.authorization).toEqual({
-      kind: "denied",
-      reason: "sender_not_allowed",
-    });
-
-    const ctx = projected.context;
-    expect(ctx).toMatchObject({
-      MessageSid: "event-1",
-      MessageSidFull: "example:event-1",
-      BodyForAgent: "agent body",
-      RawBody: "raw body",
-      CommandBody: "/status",
-      ReplyToId: "quoted-1",
-      CommandAuthorized: false,
-      ConversationLabel: "Example Room",
-      GroupSubject: "Example Room",
-      SenderE164: "+15550001111",
-      media: [
-        {
-          path: "/tmp/example.jpg",
-          contentType: "image/jpeg",
-          kind: "image",
-        },
-      ],
-    });
   });
 });
